API-1973: Wrap message body when raising event directly in SQS
When using `raise_event` to push the message to SNS, the original
message gets wrap when pushing the message to SQS.
`raise_event_in_queue` needs to wrap the message itself.

This also locks the version of aws-sdk-core to ~> 2.0, as version eventq
is not compatible to v3 of the AWS SDK yet.