Handle paging when listing topics
Calls to `#list_topics` can return paged results. When this happens the
response includes a `next_token` which we can use to request the next
page.

This PR updates `#get_topic_arn` to check for a `next_token` when making
calls to `#list_topics`. This ensures we get *all* the topics before
checking whether the given topic / event_type exists.

Note that, according to the AWS docs, calls to `#list_topics` are
throttled to 30 TPS. Therefore we call `sleep` between requests to make
sure we play nicely.