Return topic arns when registering SNS topics
Update the AWS `EventQClient#regiser_event` to return the topic arn of
the topic it registers.

Update the `#registered?` method too: if an event is already registered
then return its topic arn, otherwise return false.

Internally the `@known_event_types` has been changed from an array to
a hash. The existing topic key that was stored in the array is used as
the key, and the topic arn as the value.