Pin AWS SDK versions until we properly upgraded `eventq`
In the past month, AWS released backward incompatible changes to
`aws-sdk-core` and `aws-sdk-sqs`. We will need to upgrade `eventq` soon,
as the changes will be enforced beginning next year.

`aws-sdk-core` released 3.187.1 which changed the default response to an
empty list, instead of `nil`. Also, 3.188.0 drops support for Ruby 2.4 &
2.5, which is currently the default for `eventq` during test runs. For
more details see:
https://github.com/aws/aws-sdk-ruby/blob/version-3/gems/aws-sdk-core/CHANGELOG.md#31871-2023-11-20

`aws-sdk-sqs` released 1.66.0 which switches to using a JSON protocol we
have not adjusted to yet. So we need to use a lower version until we
upgrade. You can find the release details here:
https://github.com/aws/aws-sdk-ruby/blob/version-3/gems/aws-sdk-sqs/CHANGELOG.md#1660-2023-11-08

`aws-sdk-sns` released new versions related to new `aws-sdk-core`
releases, which align with 3.188.0 and higher. But that version drops
Ruby 2.4 support as detailed above, so need to stick with a lower
version until we upgraded to a newer Ruby version. You can find the
changelog here:
https://github.com/aws/aws-sdk-ruby/blob/version-3/gems/aws-sdk-sns/CHANGELOG.md#1680-2023-10-26