Remove default setting inside calc visibility timeout
Because the default is already set in the base queue class, it's not
necessary to have a fallback value when fetching the queue settings.

Since the new setting is now mandatory for calling the visibility
timeout class, adjust the specs to use a let-block for the queue
settings throughout the spec file to simplify it and make it easy to add
a new queue setting.