Extract calculating RabbitMQ retry delay to private method
Because the `reject_message` method became way too long, extract
calculating the retry delay to it's separate method and call it.

This refactoring also fixes that the exponential backoff in RabbitMQ was
not capped, as the check did re-calculate the condition as
`retry_attempts * queue.retry_delay` instead of using the calculated
`message_ttl` and therefore did not include exponential backoff.