Add healthchecks to wait for services being ready
In the upgraded environment with Ruby 3 booting up localstack takes too
long, so some tests relying on it fail before it becomes available.
Solution is just to add healthchecks and to wait for them to be healthy
before starting tests. Also add this for Rspec test action on GitHub to
avoid flaky test runs.