Set log level to FATAL during spec runs
Since the logger is printing messages to STDOUT, all DEBUG, INFO and
ERROR logs produced during tests are included in the test output. Since
we rely heavily on failing to process a message on purpose, we have
littered the output with ERROR logs and traces as well. By setting the
log level to FATAL during spec runs, we ensure that we only see the spec
output during test runs.