Change CI to use matrix of 2.5 to 3.2 (#106)
* CI: Move to matrix testing 2.7, 3.0, 3.2

* Bump codeclimate action

* Adjust script and source path
As we are now setting `src` as the default folder, we can just use the
container loop from the mounted `src` volume. Also adjust paths in
README and test script to work from root, rather inside the `script`
folder.

* Add healthchecks to wait for services being ready
In the upgraded environment with Ruby 3 booting up localstack takes too
long, so some tests relying on it fail before it becomes available.
Solution is just to add healthchecks and to wait for them to be healthy
before starting tests. Also add this for Rspec test action on GitHub to
avoid flaky test runs.

* Extend test matrix since we have support for Ruby 2.5+