Make it easy to log Queue settings (#114)
* Update Dockerfile to use "new" format `ENV key=value`

* Make it easy to log Queue settings

Because the Queue setup is done in the client and can be done in a
variety of ways, we cannot automatically log the settings when the Queue
is created. Instead, we provide a method to log the settings to call in
the client when appropriate.

Logging the settings will help to make sure changes to your Queue
configuration are applied and can also double check business logic
around these settings are calculating the right outcome.