Fix release process by avoiding rake (#119)
* Fix and clean up release process

When moving to trusted publishers for RubyGems, we tried the suggested
approach of using the `rubygems/release-gem@v1` workflow. However, that
is releasing via `bundle exec rake release`. That caused multiple
changes and issues.

First, we didn't have a Rakefile since we never released that way. We
added that intermittently, but is now cleaned up again. By including
`require 'bundler/gem_tasks'` we made the `rake release` command
available.

Second, the problem of this approach is that any run of `bundler` will
try resolving all dependencies and install required packages. However,
our gem is intended to be used with a variety of Ruby versions, which is
why we don't pin any gem version to leave it open to the user of the
gem. This led to conflicts as `bundler` was simply using the latest
versions across the board.

Third, you would think using the latest version is fine. However, Ruby
started to ship some gems itself by default, e.g. `set` or `openssl`.
Since `bundler` activates those first, sees the requirement on
`openssl`, it tries to install and activate the latest version of that.
But since Ruby ships it's own version and we have no version pins,
`bundler` already activated the tiny bit older `openssl` version of the
Ruby version and fails to activate the latest `openssl` version, failing
the release process.

So now we avoid using the `rubygems/release-gem@v1` action and instead
do the authentication, build and push steps ourselves, avoiding
`bundler` entirely. This way no dependencies are checked and we can
achieve maximum compatibility across Ruby versions.

* Bump up EventQ version to cut again with fixed release process